// src/utils/mocks.ts
import { TWebsocketOrder, TWebsocketOrders } from "../types/order-feed";
import { BurgerItem } from "../types/types";
import { UserType } from "../types/user-types";

export const mockUser: UserType = {
  email: "test@example.com",
  name: "Test User",
};

export const bunIngredient = {
  _id: "643d69a5c3f7b9001cfa093c",
  name: "Краторная булка N-200i",
  type: "bun",
  proteins: 80,
  fat: 24,
  carbohydrates: 53,
  calories: 420,
  price: 1255,
  image: "https://code.s3.yandex.net/react/code/bun-02.png",
  image_mobile: "https://code.s3.yandex.net/react/code/bun-02-mobile.png",
  image_large: "https://code.s3.yandex.net/react/code/bun-02-large.png",
  __v: 0,
} as BurgerItem;

export const fillingIngredient = {
  _id: "643d69a5c3f7b9001cfa0941",
  name: "Биокотлета из марсианской Магнолии",
  type: "main",
  proteins: 420,
  fat: 142,
  carbohydrates: 242,
  calories: 4242,
  price: 424,
  image: "https://code.s3.yandex.net/react/code/meat-01.png",
  image_mobile: "https://code.s3.yandex.net/react/code/meat-01-mobile.png",
  image_large: "https://code.s3.yandex.net/react/code/meat-01-large.png",
  __v: 0,
} as BurgerItem;

export const fillingIngredient2 = {
  _id: "643d69a5c3f7b9001cfa0942",
  name: "Соус Spicy-X",
  type: "sauce",
  proteins: 30,
  fat: 20,
  carbohydrates: 40,
  calories: 30,
  price: 90,
  image: "https://code.s3.yandex.net/react/code/sauce-02.png",
  image_mobile: "https://code.s3.yandex.net/react/code/sauce-02-mobile.png",
  image_large: "https://code.s3.yandex.net/react/code/sauce-02-large.png",
  __v: 0,
} as BurgerItem;
export const mockOrder: TWebsocketOrder = {
  _id: "order123",
  status: "done",
  ingredients: ["ingredient1", "ingredient2"],
  name: "testOrder",
  createdAt: "",
  updatedAt: "",
  number: 100500,
};

export const WebsocketOrders: TWebsocketOrders = {
  success: true,
  orders: [],
  total: 1,
  totalToday: 1,
};
